refactor: consolidate CLI utilities into types.ts and utils.ts
- Move all CLI types (NotFoundOptions, JsonSchema*) to src/cli/types.ts
- Move all CLI functions to src/cli/utils.ts
- Remove src/utils/ folder (unused ContentBuilder and formatters)
- Delete redundant errors.ts, formatters.ts, schema.ts from cli/
