--- src/utils/index.ts
@@ -1,21 +1,0 @@
-/**
- * Shared utilities
- * Common functions used across CLI and SDK
- */
-
-export type { ContentBuilderOptions } from "./content-builder";
-// Content building
-export { ContentBuilder, createContentBuilder } from "./content-builder";
-export type { NotFoundOptions } from "./errors";
-// Error handling
-export { formatError, handleNotFound, handleValidationError } from "./errors";
-
-// Formatting
-export {
-  definitionsToRows,
-  definitionToRow,
-  formatDuration,
-  formatFileSize,
-  formatSchemaInputs,
-  truncate,
-} from "./formatters";
